// src/dfs/generated/artifacts/operation.ts
/*
 * Copyright (c) Microsoft Corporation. All rights reserved.
 * Licensed under the MIT License. See License.txt in the project root for license information.
 *
 * Code generated by Microsoft (R) AutoRest Code Generator.
 * Changes may cause incorrect behavior and will be lost if the code is regenerated.
 */

export enum Operation {
  Service_ListFileSystems,
  Service_SetProperties,
  Service_GetProperties,
  Service_GetStatistics,
  Service_ListContainersSegment,
  Service_GetUserDelegationKey,
  Service_GetAccountInfo,
  Service_GetAccountInfoWithHead,
  Service_SubmitBatch,
  Service_FilterBlobs,
  FileSystem_Create,
  FileSystem_SetProperties,
  FileSystem_GetProperties,
  FileSystem_Delete,
  FileSystem_ListPaths,
  FileSystem_ListBlobFlatSegment,
  FileSystem_ListBlobHierarchySegment,
  Path_Create,
  Path_Update,
  Path_Lease,
  Path_Read,
  Path_GetProperties,
  Path_Delete,
  Path_SetAccessControl,
  Path_SetAccessControlRecursive,
  Path_SetProperties,
  Path_FlushData,
  Path_AppendData,
  Path_SetExpiry,
  Path_Undelete,
  Container_Create,
  Container_GetProperties,
  Container_GetPropertiesWithHead,
  Container_Delete,
  Container_SetMetadata,
  Container_GetAccessPolicy,
  Container_SetAccessPolicy,
  Container_Restore,
  Container_SubmitBatch,
  Container_FilterBlobs,
  Container_AcquireLease,
  Container_ReleaseLease,
  Container_RenewLease,
  Container_BreakLease,
  Container_ChangeLease,
  Container_GetAccountInfo,
  Container_GetAccountInfoWithHead,
  PageBlob_Create,
  PageBlob_UploadPages,
  PageBlob_ClearPages,
  PageBlob_UploadPagesFromURL,
  PageBlob_GetPageRanges,
  PageBlob_GetPageRangesDiff,
  PageBlob_Resize,
  PageBlob_UpdateSequenceNumber,
  PageBlob_CopyIncremental,
  AppendBlob_Create,
  AppendBlob_AppendBlock,
  AppendBlob_AppendBlockFromUrl,
  AppendBlob_Seal,
  BlockBlob_Upload,
  BlockBlob_PutBlobFromUrl,
  BlockBlob_StageBlock,
  BlockBlob_StageBlockFromURL,
  BlockBlob_CommitBlockList,
  BlockBlob_GetBlockList,
  Blob_Undelete,
  Blob_SetExpiry,
  Blob_SetHTTPHeaders,
  Blob_SetImmutabilityPolicy,
  Blob_DeleteImmutabilityPolicy,
  Blob_SetLegalHold,
  Blob_SetMetadata,
  Blob_AcquireLease,
  Blob_ReleaseLease,
  Blob_RenewLease,
  Blob_ChangeLease,
  Blob_BreakLease,
  Blob_CreateSnapshot,
  Blob_StartCopyFromURL,
  Blob_CopyFromURL,
  Blob_AbortCopyFromURL,
  Blob_SetTier,
  Blob_GetAccountInfo,
  Blob_GetAccountInfoWithHead,
  Blob_Query,
  Blob_GetTags,
  Blob_SetTags,
}
export default Operation;